move description-formatter to its own file and cleanup dependencies
