+++ src/runtime/plan/suggestion.ts
@@ -10,7 +10,8 @@
 
 import {assert} from '../../platform/assert-web.js';
 import {Arc} from '../arc.js';
-import {Description, DescriptionFormatter} from '../description.js';
+import {Description} from '../description.js';
+import {DescriptionFormatter} from '../description-formatter.js';
 import {Manifest} from '../manifest.js';
 import {Modality} from '../modality.js';
 import {Recipe} from '../recipe/recipe.js';
